+++ src/geefetch/cli/main.py
@@ -25,18 +25,13 @@ config_option = click.option(
 )
 @click.option("--debug", is_flag=True)
 def main(verbose, quiet, logfile, debug):
-    """The geefetch tool."""
+    """Download satellites from Google Earth Engine."""
     from .logging_implementation import logging_setup
 
     logging_setup(verbose, quiet, logfile, debug)
 
 
-@main.group()
-def download():
-    """Download satellites from Google Earth Engine."""
-
-
-@download.command()
+@main.command()
 @config_option
 def all(config):
     """Download all satellites given in the config."""
@@ -45,7 +40,7 @@ def all(config):
     download_all(config)
 
 
-@download.command()
+@main.command()
 @config_option
 @click.option(
     "--vector/--no-vector",
@@ -60,7 +55,7 @@ def gedi(config, vector):
     download_gedi(config, vector)
 
 
-@download.command()
+@main.command()
 @config_option
 def s1(config):
     """Download Sentinel-1 images."""
@@ -69,7 +64,7 @@ def s1(config):
     download_s1(config)
 
 
-@download.command()
+@main.command()
 @config_option
 def s2(config):
     """Download Sentinel-2 images."""
@@ -78,41 +73,10 @@ def s2(config):
     download_s2(config)
 
 
-@download.command()
+@main.command()
 @config_option
 def dynworld(config):
     """Download Dynamic World images."""
     from .download_implementation import download_dynworld
 
     download_dynworld(config)
-
-
-@main.group()
-def process():
-    """Pre/post processing tools for GEDI, Sentinel-1 and Sentinel-2 data."""
-
-
-@process.command()
-@config_option
-@click.option("--s1/--no-s1", default=False, help="Whether to clean s1 data.")
-@click.option("--s2/--no-s2", default=False, help="Whether to clean s2 data.")
-@click.option("--gedi/--no-gedi", default=False, help="Whether to clean gedi data.")
-@click.option(
-    "--num-threads",
-    default=8,
-    help="Number of threads to use for processing in parallel.",
-)
-def clean(config, s1, s2, gedi, num_threads):
-    """Remove corrupted .tif files."""
-    from .process_implementation import clean
-
-    clean(config, s1, s2, gedi, num_threads)
-
-
-@process.command()
-@config_option
-def merge(config):
-    """Merge GEDI geojson files."""
-    from .process_implementation import merge_geojson_dataset
-
-    merge_geojson_dataset(config)